+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-The forma is based on Keep a Changelog.
+The format is based on Keep a Changelog.
 This project uses Semantic Versioning.
 
 ---
 
-## [0.0.1] - 2026.06.01
+## [0.1.0] - 2026-06-01
+- Added main functionalities, including standard log, warning and error
+
+## [0.0.1] - 2026-06-01
 
 ### Added
 - Added first structure and few files.